refator(review) 상위 컨텍스트 추가
- 명시적인 컨텍스트를 표현하기 위해 상위 컨텍스트를 추가함
- 메뉴 : 메뉴, 메뉴그룹, 음식 / 배달음식 / 포장음식 / 매장음식 : 매장음식, 매장테이블
