--- src/main/java/kitchenpos/client/purgomalum/PurgomalumClient.java
@@ -1,5 +1,0 @@
-package kitchenpos.client.purgomalum;
-
-public interface PurgomalumClient {
-	boolean containsProfanity(String text);
-}
