--- src/main/java/kitchenpos/eatinorder/domain/EatInTable.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorder.domain;
-
-public class EatInTable {
-
-}
