--- src/main/java/kitchenpos/eatinorder/domain/EatinOrder.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorder.domain;
-
-public class EatinOrder {
-}
