--- src/main/java/kitchenpos/eatinorder/domain/OrderLineItem.java
@@ -1,4 +1,0 @@
-package kitchenpos.eatinorder.domain;
-
-public class OrderLineItem {
-}
