+++ src/main/java/kitchenpos/eatinorder/eatinorder/application/port/in/EatinOrderUseCase.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorder.eatinorder.application.port.in;
+
+public interface EatinOrderUseCase {
+
+}
