+++ src/main/java/kitchenpos/eatinorder/eatinorder/domain/OrderLineItem.java
@@ -1,0 +1,4 @@
+package kitchenpos.eatinorder.eatinorder.domain;
+
+public class OrderLineItem {
+}
