+++ src/main/java/kitchenpos/eatinorder/eatintable/domain/EatInTable.java
@@ -1,0 +1,5 @@
+package kitchenpos.eatinorder.eatintable.domain;
+
+public class EatInTable {
+
+}
