--- src/main/java/kitchenpos/menu/adapter/in/ProductRestController.java
@@ -1,10 +1,0 @@
-package kitchenpos.menu.adapter.in;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RequestMapping("/api/products")
-@RestController
-public class ProductRestController {
-
-}
