--- src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,7 +1,0 @@
-package kitchenpos.menu.application;
-
-import kitchenpos.menu.application.port.in.MenuGroupUseCase;
-
-public class MenuGroupService implements MenuGroupUseCase {
-
-}
