--- src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,7 +1,0 @@
-package kitchenpos.menu.application;
-
-import kitchenpos.menu.application.port.in.MenuUseCase;
-
-public class MenuService implements MenuUseCase {
-
-}
