--- src/main/java/kitchenpos/menu/application/port/in/MenuGroupUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.application.port.in;
-
-public interface MenuGroupUseCase {
-
-}
