--- src/main/java/kitchenpos/menu/application/port/in/ProductUseCase.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.application.port.in;
-
-public interface ProductUseCase {
-
-}
