--- src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,20 +1,0 @@
-package kitchenpos.menu.domain;
-
-import java.math.BigDecimal;
-import java.util.List;
-import java.util.UUID;
-
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.ForeignKey;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.persistence.Transient;
-
-public class Menu {
-
-}
