--- src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,5 +1,0 @@
-package kitchenpos.menu.domain;
-
-public class MenuProduct {
-
-}
