--- src/main/java/kitchenpos/menu/domain/Product.java
@@ -1,13 +1,0 @@
-package kitchenpos.menu.domain;
-
-import java.math.BigDecimal;
-import java.util.UUID;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-public class Product {
-
-}
