+++ src/main/java/kitchenpos/menu/menu/application/MenuService.java
@@ -1,0 +1,7 @@
+package kitchenpos.menu.menu.application;
+
+import kitchenpos.menu.menu.application.port.in.MenuUseCase;
+
+public class MenuService implements MenuUseCase {
+
+}
