+++ src/main/java/kitchenpos/menu/menu/application/port/in/MenuUseCase.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.menu.application.port.in;
+
+public interface MenuUseCase {
+
+}
