+++ src/main/java/kitchenpos/menu/menu/domain/Menu.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.menu.domain;
+
+public class Menu {
+
+}
