+++ src/main/java/kitchenpos/menu/menugroup/adapter/in/MenuGroupRestController.java
@@ -1,0 +1,10 @@
+package kitchenpos.menu.menugroup.adapter.in;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RequestMapping("/api/menu-groups")
+@RestController
+public class MenuGroupRestController {
+
+}
