+++ src/main/java/kitchenpos/menu/menugroup/domain/MenuGroup.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.menugroup.domain;
+
+public class MenuGroup {
+
+}
