+++ src/main/java/kitchenpos/menu/product/application/port/in/ProductUseCase.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.product.application.port.in;
+
+public interface ProductUseCase {
+
+}
