+++ src/main/java/kitchenpos/menu/product/domain/Product.java
@@ -1,0 +1,5 @@
+package kitchenpos.menu.product.domain;
+
+public class Product {
+
+}
